// src/api/model/metric.ts
/**
 * Onepanel
 * Onepanel API
 *
 * The version of the OpenAPI document: 0.12.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Metric { 
    name?: string;
    value?: number;
    format?: string;
}